requirements.txt
# Web framework
fastapi==0.104.1
uvicorn==0.24.0

# MongoDB
motor==3.3.2
pymongo==4.6.0

# Validación
pydantic==2.5.0
pydantic-settings==2.1.0

# Auth
python-jose[cryptography]==3.3.0

#Cliente HTTP para verificacion de google
httpx==0.25.2

# Env
python-dotenv==1.0.0

# Multipart 
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
